pyproject.toml
[project]
name = "rag-practice"
version = "0.1.0"
description = "Learning RAG (Retrieval-Augmented Generation) with LangChain"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-chroma>=0.1",
    "langchain-google-genai>=2.0",
    "langchain-huggingface>=0.1",
    "chromadb>=0.5",
    "sentence-transformers>=3.0",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.scripts]
rag = "main:main"
